Commit 10/06/2020 5:30  Commit pre entrega

+++ MendoLopez-Carlos-general-Apuntate/Controllers/ApuntateController.cs
@@ -35,40 +35,171 @@ namespace MendoLopez_Carlos_general_Apuntate.Controllers
         }
 
         [HttpGet]
-        public ActionResult<List<Evento>> Get() =>
-            _apuntateService.Get();
+        public ActionResult<List<Evento>> GetAllEventos() =>
+            _apuntateService.Get(); //Obtener todos los eventos
 
         [HttpGet("{id}", Name="GetEvento")]
-        public ActionResult<Evento> Get(string id){
+        public ActionResult<Evento> GetEvento(string id){ //Obtener informacion de un evento concreto
+            if(id == null || id == ""){
+                return BadRequest("El campo id debe contener informacion valida");
+            }
             var evento = _apuntateService.Get(id);
             if(evento == null){
-                return NotFound();
+                return NotFound("La informacion del evento debe ser valida");
             }
             return evento;
         }
 
+        [HttpPatch("{id}/{email}")]
+        public IActionResult reservarTurno( //Usada por profesores y alumnos
+            //Se reserva el turno de un evento
+            string id,
+            [FromQuery] string turno,
+            string email)
+            {
+                if(id == null || id == ""){
+                    return BadRequest("El id debe contener informacion valida");
+                }
+
+                if(turno == null || turno == ""){
+                    return BadRequest("El turno debe contener informacion valida");
+                }
+
+                if(email == null || email ==""){
+                    return BadRequest("El email debe contener informacion valida");
+                }
+
+                var evento = _apuntateService.Get(id);
+                if(evento == null){
+                    return NotFound("El evento no se ha encontrado");
+                }
+
+                foreach (Turno turn in evento.grupos){
+                    if(turn.turno == turno){ //Si es el mismo,vemos si se ha asignado, y si no, se asigna.En caso contrario, a la lista de espera
+                        if(turn.actual == null && !turn.listaEspera.Contains(email)){
+                            turn.actual = email;
+                        } else {
+                            if(!turn.listaEspera.Contains(email)){
+                                turn.listaEspera.Add(email);
+                            } else {
+                                return BadRequest("Un usuario no se puede registrar dos veces en el mismo turno");
+                            }
+                        }
+                    }
+                }
+
+                //TRATAMIENTO CON COLAS
+                actualizarTarea(evento,email,VALOR_CONTROL_COMPLETADA); //El usuario ha completado el registro
+                _apuntateService.Update(evento.Id,evento);
+                return NoContent();
+
+
+            }
+
+        [HttpPut("{id}/{email}")]
+        public IActionResult cancelarTurno( //Usada por profesores y  alumnos
+            //Se cancela la reserva de un turno
+            string id,
+            [FromQuery] string turno,
+            string email)
+            {
+                if(id == null || id == ""){
+                    return BadRequest("El id debe contener informacion valida");
+                }
+
+                if(turno == null || turno == ""){
+                    return BadRequest("El turno debe contener informacion valida");
+                }
+
+                if(email == null || email ==""){
+                    return BadRequest("El email debe contener informacion valida");
+                }
+                bool resul = isAlumno(email).Result;
+                if(!resul){
+                    return Unauthorized("Un profesor no puede cancelar un turno");
+                }
+
+                var evento = _apuntateService.Get(id);
+                if(evento == null){
+                    return NotFound("El evento no se ha encontrado");
+                }
+
+                foreach (Turno turn in evento.grupos){
+                    if(turn.turno == turno){ //Si es el mismo,vemos si se ha asignado, y si no, se asigna.En caso contrario, a la lista de espera
+                        if(turn.actual == email){
+                            var nuevo = turn.listaEspera.ToArray()[0];
+                            if(nuevo != null){
+                                turn.actual = nuevo;
+                                turn.listaEspera.Remove(nuevo);
+                            } else {
+                                turn.actual = null;
+                            }
+                        } else { //Si no esta en el turno principal, se borra de la lista de espera
+                            turn.listaEspera.Remove(email);
+                        }
+                }
+                }
+                //TRATAMIENTO CON COLAS
+                //actualizarTarea(evento,email,VALOR_CONTROL_COMPLETADA); //El usuario ha completado el registro
+                _apuntateService.Update(evento.Id,evento);
+                return NoContent();
+
+
+            }
+
         [HttpPost]
-        public ActionResult<Evento> Create(Evento evento,
+        public ActionResult<Evento> CreateEvento(Evento evento,
             [FromQuery] string email){
-                Console.WriteLine(email);
+            //Crear evento
+                
+            if(evento == null){
+                return BadRequest("El evento debe contener informacion valida");
+            }
+
+            if(email == null || email == ""){
+                return BadRequest("El email debe contener informacion valida");
+            }
+
+
             bool resul = isAlumno(email).Result;
-            if(resul){
+            if(!resul){
                 Evento a = _apuntateService.Create(evento);
                 actualizarTarea(a,email,VALOR_CONTROL_PENDIENTE);
                 return CreatedAtRoute("GetEvento",new {id = evento.Id.ToString()},evento);
             } else {
-                return NotFound();
+                return Unauthorized("Un alumno no puede crear un evento");
             }
+
+            
         }
 
+
+
         [HttpPut("{id}")]
-        public IActionResult Update(
+        public IActionResult updateEvento(
+            //Actualizar evento
             string id, 
             Evento eventoMod,
             [FromQuery] string email){
+
+            if(id == null||id==""){
+                return BadRequest("La informacion del id debe ser valida");
+            }    
+
+            if(eventoMod == null){
+                return BadRequest("La informacion del evento debe ser valida");
+            }
+
+            if(email == null || email == ""){
+                return BadRequest("La informacion del email debe ser valida");
+            }
             var evento = _apuntateService.Get(id);
             if(evento == null){
-                return NotFound();
+                return NotFound("No se ha encontrado el evento");
+            }
+            bool resul = isAlumno(email).Result;
+            if(resul){
+                return Unauthorized("Un alumno no puede modificar un evento");
             }
 
             _apuntateService.Update(id,eventoMod);
@@ -76,21 +207,41 @@ namespace MendoLopez_Carlos_general_Apuntate.Controllers
         }
 
         [HttpDelete("{id}")]
-        public IActionResult Delete(
+        public IActionResult deleteEvento(
+            //Eliminar evento
             string id,
             [FromQuery] string email){
+
+            if(id == null || id == ""){
+                return BadRequest("El id debe contener informacion valida");
+            }
+
+            if(email == null || email==""){
+                return BadRequest("El email debe ser valido");
+            }
+
+            bool resul = isAlumno(email).Result;
+            if(resul){
+                return Unauthorized("Un alumno no puede eliminar un evento");
+            }
+
+
             var evento = _apuntateService.Get(id);
 
             if(evento == null){
-                return NotFound();
+                return NotFound("No se ha encontrado el evento");
             }
 
+            actualizarTarea(evento,email,VALOR_CONTROL_ELIMINAR);
+
             _apuntateService.Remove(evento.Id);
+
+
 
             return NoContent();
         }
 
-        private async Task<bool> isAlumno(string email){
+        private async Task<bool> isAlumno(string email){ //Comprobar si un email pertenece a un alumno
             string requestUrl = "http://localhost:8086/api/usuarios/" + email;
             Console.WriteLine(requestUrl);
             using var client = new HttpClient();
@@ -103,10 +254,10 @@ namespace MendoLopez_Carlos_general_Apuntate.Controllers
             } else {
                 return false;
             }
-            //COMPLETAR
+
         }
 
-        private void actualizarTarea(Evento e, string email, int modificador){
+        private void actualizarTarea(Evento e, string email, int modificador){ //Enviar mensajes a la cola de mensajes
             var factory = new ConnectionFactory() { Uri = new Uri("amqp://cikqdgnl:IpTo-xh4cN-UU2Aa0JX2nf1AN3chkLB4@squid.rmq.cloudamqp.com/cikqdgnl") };
             var connection = factory.CreateConnection();
             var channel = connection.CreateModel();
@@ -142,7 +293,7 @@ namespace MendoLopez_Carlos_general_Apuntate.Controllers
             connection.Close();
         }
 
-        private string formarMensaje(Evento evento, string email, int modificador){
+        private string formarMensaje(Evento evento, string email, int modificador){ //Metodo de creacion para los mensajes de la cola
 
             string mensaje = null;
             switch (modificador)
@@ -159,6 +310,7 @@ namespace MendoLopez_Carlos_general_Apuntate.Controllers
 
                 case VALOR_CONTROL_ELIMINAR:
                     MensajeEliminar eliminar = new MensajeEliminar(modificador, evento.Id);
+                    mensaje = JsonSerializer.Serialize(eliminar);
                     break;
                 
                 default:
